+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Upgrade to stable Rust 1.84.0.
-- Upgrade widestring dependency to 1.1.0
+- Upgrade widestring dependency to 1.1.0.
+- Replace deprecated winapi dependency with windows-sys.
 
 ### Security